javadocs and minor updates

+++ src/main/java/com/ashisht/mini_bank/entity/OperationType.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 public class OperationType {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private Long operationTypeId;
+    private Integer operationTypeId;
     private String description;
 }
