+++ src/main/java/com/ashisht/mini_bank/repository/OperationTypeRepo.java
@@ -3,6 +3,6 @@ package com.ashisht.mini_bank.repository;
 import com.ashisht.mini_bank.entity.OperationType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OperationTypeRepo extends JpaRepository<OperationType, Long> {
+public interface OperationTypeRepo extends JpaRepository<OperationType, Integer> {
     boolean existsByDescription(String description);
 }
